+++ db/src/main/java/io/nuls/db/entity/TransactionPo.java
@@ -14,7 +14,7 @@ public class TransactionPo {
 
     private Long createTime;
 
-    private Integer blockHeight;
+    private Long blockHeight;
 
     private String blockHash;
 
@@ -47,11 +47,11 @@ public class TransactionPo {
         this.createTime = createTime;
     }
 
-    public Integer getBlockHeight() {
+    public Long getBlockHeight() {
         return blockHeight;
     }
 
-    public void setBlockHeight(Integer blockHeight) {
+    public void setBlockHeight(Long blockHeight) {
         this.blockHeight = blockHeight;
     }
